+++ src/app/Controllers/Base/RequestValidationController.py
@@ -29,26 +29,6 @@ class RequestValidationController(ObjectValidationController):
         else:
             return org_name
 
-    def validate_create_task_request(self, request_body: dict, **kwargs) -> dict:
-        """Validates a task request body
-        :param request_body:    The request body from the create task request
-        :return:                Response if the request body contains invalid values, or the TaskRequest dataclass
-        """
-        return {
-            "type": self.check_task_type_id(task_type_id=request_body.get("type_id")),
-            "description": self.check_optional_str(request_body.get("description"), "description"),
-            "time_estimate": self.check_optional_int(
-                request_body.get("time_estimate"), "time_estimate", allow_negative=True
-            ),
-            "scheduled_for": self.check_optional_date(request_body.get("scheduled_for"), "scheduled_for"),
-            "scheduled_notification_period": self.check_optional_int(
-                request_body.get("scheduled_notification_period"), "scheduled_notification_period"
-            ),
-            "assignee": self.check_task_assignee(request_body.get("assignee"), **kwargs),
-            "priority": self.check_task_priority(request_body.get("priority")),
-            "labels": self.check_task_labels(request_body.get("labels", []), kwargs["req_user"].org_id),
-        }
-
     def validate_create_task_type_request(
         self, request_body: dict, **kwargs
     ) -> typing.Tuple[dict, typing.Optional[TaskType]]:
@@ -246,25 +226,6 @@ class RequestValidationController(ObjectValidationController):
                 raise ValidationError("That organisation name already exists.")
 
         return org_name
-
-    def validate_update_task_request(self, request_body: dict, **kwargs) -> dict:
-        """Validates an update task request"""
-        task = self.check_task_id(request_body.get("id"), kwargs["req_user"].org_id)
-        return {
-            "task": task,
-            "description": self.check_optional_str(request_body.get("description"), "description"),
-            "status": self.check_task_status(request_body.get("status")),
-            "time_estimate": self.check_optional_int(
-                request_body.get("time_estimate"), "time_estimate", allow_negative=True
-            ),
-            "scheduled_for": self.check_optional_date(request_body.get("scheduled_for"), "scheduled_for"),
-            "scheduled_notification_period": self.check_optional_int(
-                request_body.get("scheduled_notification_period"), "scheduled_notification_period"
-            ),
-            "assignee": self.check_task_assignee(request_body.get("assignee"), **kwargs),
-            "priority": self.check_task_priority(request_body.get("priority")),
-            "labels": self.check_task_labels(request_body.get("labels", []), kwargs["req_user"].org_id),
-        }
 
     def validate_update_task_type_request(self, org_id: int, request_body: dict) -> typing.Tuple[TaskType, dict, list]:
         """ Validates an update task type request body """