+++ src/app/Extensions/Errors.py
@@ -68,3 +68,21 @@ class ResourceNotFoundError(Exception):
         rv = dict(self.payload or ())
         rv["msg"] = self.message
         return rv
+
+
+class InternalServerError(Exception):
+    """Error when a resource is not found"""
+
+    status_code = 500
+
+    def __init__(self, message, status_code=None, payload=None):
+        Exception.__init__(self)
+        self.message = message
+        if status_code is not None:
+            self.status_code = status_code
+        self.payload = payload
+
+    def to_dict(self):
+        rv = dict(self.payload or ())
+        rv["msg"] = self.message
+        return rv