+++ src/app/Models/Subscription.py
@@ -1,44 +1,21 @@
-from app import logger, subscription_api
+from os import getenv
+
+import requests
+from flask import current_app
+
+from app.Extensions.Errors import InternalServerError
 
 
 class Subscription(object):
     def __init__(self, subscription_id: str):
         self._subscription_id = subscription_id
 
-        sub_request = subscription_api.get_subscription_meta(self._subscription_id)
+        subscription_meta = self._get_subscription_meta(subscription_id)
 
-        self._status = sub_request["status"]
-        self._metadata = sub_request["meta_data"]
+        self._status = subscription_meta["status"]
+        self._metadata = subscription_meta["meta_data"]
         self._unusable_statuses = ["future", "paused", "cancelled"]
         self._usable_statuses = ["active", "non_renewing"]
-
-    def max_users(self) -> int:
-        """The maximum users that can be created on this plan"""
-        if self._status == "in_trial":
-            return -1
-        if self._status in self._unusable_statuses:
-            return 0
-        if self._status in self._usable_statuses:
-            ret = self._metadata.get("max_users")
-            if ret is None:
-                logger.warning(f"Missing max_users metadata for {self._subscription_id}")
-                return -1
-            else:
-                return ret
-
-    def can_get_reports(self) -> bool:
-        """Can the subscription get reports?"""
-        if self._status == "in_trial":
-            return True
-        if self._status in self._unusable_statuses:
-            return False
-        if self._status in self._usable_statuses:
-            ret = self._metadata.get("view_reports_page")
-            if ret is None:
-                logger.warning(f"Missing view_reports_page metadata for {self._subscription_id}")
-                return True
-            else:
-                return ret
 
     def task_activity_log_history(self) -> int:
         """The amount of days of task activity log that can be viewed"""
@@ -49,26 +26,12 @@ class Subscription(object):
         if self._status in self._usable_statuses:
             ret = self._metadata.get("task_activity_log_history")
             if ret is None:
-                logger.warning(f"Missing task_activity_log_history metadata for {self._subscription_id}")
+                current_app.logger.warning(f"Missing task_activity_log_history metadata for {self._subscription_id}")
                 return -1
             else:
                 return ret
 
-    def can_search_dashboard(self) -> bool:
-        """Can the subscription search the dashboard"""
-        if self._status == "in_trial":
-            return True
-        if self._status in self._unusable_statuses:
-            return False
-        if self._status in self._usable_statuses:
-            ret = self._metadata.get("searchable_dashboard")
-            if ret is None:
-                logger.warning(f"Missing searchable_dashboard metadata for {self._subscription_id}")
-                return True
-            else:
-                return ret
-
-    def can_view_user_activity(self) -> bool:
+    def user_activity_log_history(self) -> bool:
         """Can the subscription view user activity logs"""
         if self._status == "in_trial":
             return True
@@ -77,7 +40,26 @@ class Subscription(object):
         if self._status in self._usable_statuses:
             ret = self._metadata.get("view_user_activity")
             if ret is None:
-                logger.warning(f"Missing view_user_activity metadata for {self._subscription_id}")
+                current_app.logger.warning(f"Missing view_user_activity metadata for {self._subscription_id}")
                 return True
             else:
                 return ret
+
+    def _get_subscription_meta(self, subscription_id) -> dict:
+        if getenv("MOCK_SERVICES"):
+            return {"status": "in_trial", "meta_data": {}}
+        try:
+            r = requests.get(
+                url=f"{self.url}/subscription/{subscription_id}/quantity",
+                headers={"Authorization": f"Bearer {self.create_sa_token()}"},
+                timeout=10,
+            )
+            if r.status_code != 200:
+                current_app.logger.error(f"There was an error getting the subscription quantity for {subscription_id}")
+                raise InternalServerError("Something went wrong getting details about your subscription!")
+            else:
+                return r.json()
+
+        except requests.exceptions.RequestException:
+            current_app.logger.error(f"There was an error getting the subscription quantity for {subscription_id}")
+            raise InternalServerError("Something went wrong getting details about your subscription!")